refactor losses to fv3fit.emulation.losses
